requirements.txt
elasticsearch==9.0.2
fastapi==0.115.12
huggingface_hub==0.29.2
Pillow==11.2.1
python-dotenv==1.1.0
sentence_transformers==4.1.0
torch==2.7.1
torchvision==0.22.1
transformers==4.49.0
